requirements.txt
django-model-utils
django
djangorestframe